refactor: network 패키지를 remote로 이름 변경 및 관련 임포트 수정
