+++ app/src/main/java/woowacourse/shopping/data/remote/HttpProductServer.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.network
+package woowacourse.shopping.data.remote
 
 import okhttp3.mockwebserver.MockWebServer
