+++ app/src/main/java/woowacourse/shopping/data/remote/NetworkManager.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.network
+package woowacourse.shopping.data.remote
 
 import android.content.Context
 import android.net.ConnectivityManager
