+++ app/src/main/java/woowacourse/shopping/data/remote/NetworkObserver.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.network
+package woowacourse.shopping.data.remote
 
 import kotlinx.coroutines.flow.Flow
