+++ app/src/test/java/woowacourse/shopping/ui/shopping/ShoppingViewModelTest.kt
@@ -15,14 +15,13 @@ import woowacourse.shopping.data.localdb.dao.CartItemDao
 import woowacourse.shopping.data.localdb.dao.RecentItemDao
 import woowacourse.shopping.data.localdb.entity.CartItemEntity
 import woowacourse.shopping.data.localdb.entity.RecentItemEntity
-import woowacourse.shopping.data.network.NetworkObserver
+import woowacourse.shopping.data.remote.NetworkObserver
 import woowacourse.shopping.data.repository.CartRepository
 import woowacourse.shopping.data.repository.ProductRepository
 import woowacourse.shopping.data.repository.RecentItemRepository
 import woowacourse.shopping.model.Money
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.model.ProductName
-import java.io.IOException
 
 @OptIn(ExperimentalCoroutinesApi::class)
 class ShoppingViewModelTest {
